Re-factor code nicely
