+++ public/js/components/AtomEdit/CustomEditors/CTAEditor.js
@@ -1,20 +1,8 @@
 import React, { PropTypes } from 'react';
-import _get from 'lodash.get';
-import _set from 'lodash.set';
 
-
+import {ManagedForm, ManagedField} from '../../ManagedEditor';
 import FormFieldTextInput from '../../FormFields/FormFieldTextInput';
-
-const isValidHttpsUrl = (value) => {
-  const stringValue = typeof value === 'string' ? value : '';
-
-  if (stringValue.match(/^(https?:\/\/)?([\da-z\.-]+)\.([a-z\.]{2,6})([\/\w \.-]*)*\/?$/)) {
-    return true;
-  } else {
-    return "Not a HTTPS url";
-  }
-};
-
+import {isHttpsUrl} from '../../../util/validators';
 
 export class CTAEditor extends React.Component {
 
@@ -27,45 +15,16 @@ export class CTAEditor extends React.Component {
 
   render () {
 
-    const ManagedAtomEditorForm = (props) => {
-      const hydratedChildren = React.Children.map(props.children, (child) => {
-        return React.cloneElement(child, {
-          atom: props.atom,
-          updateAtom: props.updateAtom
-        });
-      });
-      return <div>{hydratedChildren}</div>;
-    };
-
-    const ManagedField = (props) => {
-
-      const updateFn = (e) => {
-        const newAtom = Object.assign({}, props.atom);
-        _set(newAtom, props.fieldLocation, e.target.value);
-        props.updateAtom(newAtom);
-      };
-
-      const hydratedChildren = React.Children.map(props.children, (child) => {
-        return React.cloneElement(child, {
-          fieldName: props.name,
-          fieldLabel: props.name,
-          fieldValue: _get(props.atom, props.fieldLocation),
-          onUpdateField: updateFn
-        });
-      });
-      return <div>{hydratedChildren}</div>;
-    };
-
     return (
       <div className="editor editor-cta">
-        <ManagedAtomEditorForm atom={this.props.atom} updateAtom={this.props.onUpdate}>
+        <ManagedForm atom={this.props.atom} updateAtom={this.props.onUpdate}>
           <ManagedField fieldLocation="data.url" name="Link Url" isRequired={true}>
             <FormFieldTextInput/>
           </ManagedField>
           <ManagedField fieldLocation="data.btnText" name="Button Text">
             <FormFieldTextInput />
           </ManagedField>
-          <ManagedField fieldLocation="data.backgroundImage" name="Background Image Url" customValidation={isValidHttpsUrl}>
+          <ManagedField fieldLocation="data.backgroundImage" name="Background Image Url" customValidation={isHttpsUrl}>
             <FormFieldTextInput />
           </ManagedField>
           <ManagedField fieldLocation="data.label" name="Background Text">
@@ -74,7 +33,7 @@ export class CTAEditor extends React.Component {
           <ManagedField fieldLocation="data.trackingcode" name="Tracking Code">
             <FormFieldTextInput />
           </ManagedField>
-        </ManagedAtomEditorForm>
+        </ManagedForm>
       </div>
     );
   }
