+++ public/js/components/ManagedEditor/index.js
@@ -1,0 +1,2 @@
+export {ManagedField} from './ManagedEditorField';
+export {ManagedForm} from './ManagedEditorForm';
